feat: :lipstick: rename statusColor to statusStyle and add icon selection function in it
