+++ src/components/Drawer/ClusterDrawer/index.tsx
@@ -23,7 +23,7 @@ import {
   poiStatusTypeMessageKeys,
   poiStatusValueMessageKeys,
 } from "../../../constants/model/poi";
-import statusColor from "../../../constants/statusColor";
+import { statusColor } from "../../../constants/statusStyle";
 import { getDownloadURL, getStorage, ref } from "firebase/storage";
 import { firebaseApp } from "../../../utils/firebase";
